# gin-learn/06-upload-file-ddd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于领域驱动设计的文件上传系统API，支持单文件上传、多文件上传、文件夹上传及大文件分片上传",
        "title": "文件上传系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "开发团队",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/upload/chunk": {
            "post": {
                "description": "上传单个文件分片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传"
                ],
                "summary": "上传文件分片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上传ID",
                        "name": "upload_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分片索引，从0开始",
                        "name": "index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "总分片数",
                        "name": "total",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "分片数据",
                        "name": "chunk",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分片上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.ChunkUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/upload/chunk/init": {
            "post": {
                "description": "开始大文件分片上传前的初始化操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传"
                ],
                "summary": "初始化分片上传",
                "parameters": [
                    {
                        "description": "初始化参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ChunkInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "初始化成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.ChunkInitResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/upload/folder": {
            "post": {
                "description": "上传完整文件夹，保留目录结构",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传"
                ],
                "summary": "文件夹上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件夹中的文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件夹名称",
                        "name": "folderName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.FolderUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/upload/merge": {
            "post": {
                "description": "所有分片上传完成后，合并为完整文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传"
                ],
                "summary": "合并文件分片",
                "parameters": [
                    {
                        "description": "合并参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ChunkMergeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "合并成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.ChunkMergeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或分片不完整",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/upload/multiple": {
            "post": {
                "description": "一次上传多个文件",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传"
                ],
                "summary": "多文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的多个文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.MultipleUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/upload/single": {
            "post": {
                "description": "上传单个文件",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传"
                ],
                "summary": "单文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.SingleUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.ChunkInitRequest": {
            "type": "object",
            "properties": {
                "fileName": {
                    "type": "string",
                    "example": "large-video.mp4"
                },
                "fileSize": {
                    "type": "integer",
                    "example": 104857600
                },
                "mimeType": {
                    "type": "string",
                    "example": "video/mp4"
                },
                "totalChunks": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "http.ChunkInitResponse": {
            "type": "object",
            "properties": {
                "upload_id": {
                    "type": "string",
                    "example": "c0e0d7b1-9eb1-4c70-a9c5-69f7ca214285"
                }
            }
        },
        "http.ChunkMergeRequest": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string",
                    "example": "video/mp4"
                },
                "file_name": {
                    "type": "string",
                    "example": "large-video.mp4"
                },
                "total_chunks": {
                    "type": "integer",
                    "example": 20
                },
                "upload_id": {
                    "type": "string",
                    "example": "c0e0d7b1-9eb1-4c70-a9c5-69f7ca214285"
                }
            }
        },
        "http.ChunkMergeResponse": {
            "type": "object",
            "properties": {
                "file": {
                    "$ref": "#/definitions/http.FileResponse"
                }
            }
        },
        "http.ChunkUploadResponse": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer",
                    "example": 3
                },
                "total": {
                    "type": "integer",
                    "example": 20
                },
                "upload_id": {
                    "type": "string",
                    "example": "c0e0d7b1-9eb1-4c70-a9c5-69f7ca214285"
                }
            }
        },
        "http.FileResponse": {
            "type": "object",
            "properties": {
                "mimeType": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "name": {
                    "type": "string",
                    "example": "example.jpg"
                },
                "path": {
                    "type": "string",
                    "example": "/uploads/2023/04/example.jpg"
                },
                "size": {
                    "type": "integer",
                    "example": 1024000
                },
                "url": {
                    "type": "string",
                    "example": "http://localhost:8080/uploads/2023/04/example.jpg"
                }
            }
        },
        "http.FolderUploadResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.FileResponse"
                    }
                },
                "folderName": {
                    "type": "string",
                    "example": "my-project"
                },
                "folderPath": {
                    "type": "string",
                    "example": "/uploads/folders/my-project"
                },
                "total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "http.MultipleUploadResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.FileResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "http.SingleUploadResponse": {
            "type": "object",
            "properties": {
                "file": {
                    "$ref": "#/definitions/http.FileResponse"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}